scripts/requirements.txt
numpy
pandas
mysql-connector-python
python-dotenv
openpyxl